refactor(search): Simplifique SearchScreen extrayendo componentes e implementando un enlace personalizado para la lógica de búsqueda
